s/SmartPointer/CoerceReferent/g
move derive_smart_pointer into removed set
